requirements.txt
beautifulsoup4==4.13.4
requests==2.32.4
pydantic-settings==2.10.1
python-dotenv==1.1.1